map statically component type to its corresponding enum. Leverage this in GetComponent()
