+++ src/GameObject.h
@@ -7,9 +7,6 @@
 #include <memory>
 
 #include "IGameObject.h"
-#include "TransformComponent.h"
-#include "RendererComponent.h"
-#include "TextureComponent.h"
 
 namespace GameEngine {
 
@@ -36,6 +33,11 @@ namespace GameEngine {
         void AddComponent(GameObjectComponentType type, std::any arg) final; // cannot be overridden
         std::shared_ptr<IGameObjectComponent> GetComponent(GameObjectComponentType type) const override;
 
+        /// template GetComponent
+        template<typename T>
+        std::shared_ptr<T> GetComponent() const {
+            return std::dynamic_pointer_cast<T>(GetComponent(ComponentTypes<T>::type));
+        }
         std::shared_ptr<TransformComponent> GetTransform() const;
         std::shared_ptr<RendererComponent> GetRenderer() const;
         std::shared_ptr<TextureComponent> GetTexture() const;
