+++ src/Types.h
@@ -5,25 +5,6 @@
 
 namespace GameEngine {
 
-/// These types correspond to components' priority
-/// and the order in which they're processed
-#define _GAME_OBJECT_COMPONENT_TYPES_ \
-    GOC_X(TRANSFORM) \
-    GOC_X(RENDERER) \
-    GOC_X(TEXTURE)
-
-#define GOC_X(name) name,
-    enum class GameObjectComponentType : unsigned int {
-        _GAME_OBJECT_COMPONENT_TYPES_
-    };
-#undef GOC_X
-#define GOC_X(name) #name,
-    constexpr const char* GameObjectComponentNames[] = {
-            _GAME_OBJECT_COMPONENT_TYPES_
-    };
-#undef GOC_X
-#undef _GAME_OBJECT_COMPONENT_TYPES_
-
     // Copy of SDL_Rect
     struct Rect {
         int x = 0;
